feat: criando pasta services e repositories
