+++ backend/src/routes/brands.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import allBrandsController from "../controllers/brands/allBrands";
+import allBrandsController from "../controllers/brands/all-brands-controller";
 
 const routeBrands = Router();
